Level’s ‘map’ is now an HTMLCanvasElement
Putting an ‘ImageData’ directly to Crafty’s canvas messes its context’s
dirty rectangles and posterior drawing requests sometimes get ignored as
a consequence.  Moreover, we need to track the position of the view port
and adjust the position where to put the image data if we want to scroll
the level too.

This commit changes the ‘map’ to be an additional canvas that is used as
an image source for the ‘drawImage’ that is used instead of
‘putImageData’.  ‘drawImage’ not only keeps the dirty rectangles
updated, but takes into account the current context’s transformation and
thus draws the image to the correct position that Crafty already set up
for us.  I suspect that this method is also slightly faster—due to the
fact that the canvas no longer needs to convert the pixel format, if it
is different to the RGBA array returned by ‘createImageData’—but I did
not test for performance.